create branch for 1.01 compatibility fixes

+++ lib-src/zipios++/zipios++/zipios_defs.h
@@ -1,13 +1,10 @@
 #ifndef ZIPIOS_DEFS_H
 #define ZIPIOS_DEFS_H
 
-#include "zipios++/zipios-config.h"
-#include <stdint.h>
-
 namespace zipios {
 
-typedef uint16_t uint16 ;
-typedef uint32_t uint32 ;
+typedef unsigned short uint16 ;
+typedef unsigned long  uint32 ;
 
 } // namespace
